requirements.txt
openpyxl
django
gunicorn
pandas
python-dotenv
mysqlclient
boto3
pip install mysql-connector-python